MomInMe/requirements.txt
asgiref==3.8.1
dj-rest-auth==6.0.0
Django==5.0.6
django-allauth==0.63.3
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-camel-case==1.4.2
pillow==10.4.0
sqlparse==0.5.0
tzdata==2024.1
